STUDY-07 | Moving ExampleSentence logic to ExampleSentenceService
